refactor(moment): extraire ParticipantAvatarStack pour dédupliquer la rangée d'avatars
Même pattern qu'en M1 Circle (MemberAvatarStack) : la rangée d'avatars
superposés avec tooltip au survol était dupliquée entre les branches
isAuthenticated=true (avec MomentRegistrationsDialog cliquable) et
isAuthenticated=false (div statique) dans moment-detail-view.tsx.
~28 lignes × 2 branches = ~56 lignes répétées.

Extraction en composant dédié src/components/moments/participant-avatar-stack.tsx.
Les domaines restent séparés de MemberAvatarStack côté Circle car le
type d'entrée diffère (RegistrationWithUser vs CircleMemberWithUser) et
les deux contextes peuvent évoluer indépendamment.
